+++ src/ports/commands.ts
@@ -6,7 +6,6 @@ import type {
   DocContent,
   DocMeta,
   GlobalCaptureSettings,
-  ImageAsset,
   LocationDescriptor,
   LocationId,
   MarkdownProfile,
@@ -46,7 +45,6 @@ import type {
   GlobalCaptureValidateShortcutParams,
   ImageDeleteParams,
   ImageImportParams,
-  ImageListParams,
   LocParams,
   PersistedSidebarTreeState,
   PersistedStyleCheckSettings,
@@ -478,16 +476,12 @@ export function appVersionGet(...[onOk, onErr]: LocParams<string>): Cmd {
   return invokeCmd<string>("app_version_get", {}, onOk, onErr);
 }
 
-export function imageImport(...[locationId, sourcePath, onOk, onErr]: ImageImportParams<string>): Cmd {
-  return invokeCmd<string>("image_import", { locationId, sourcePath }, onOk, onErr);
+export function imageImport(...[locationId, sourcePath, targetDir, onOk, onErr]: ImageImportParams<string>): Cmd {
+  return invokeCmd<string>("image_import", { locationId, sourcePath, targetDir }, onOk, onErr);
 }
 
 export function imageDelete(...[locationId, assetPath, onOk, onErr]: ImageDeleteParams<boolean>): Cmd {
   return invokeCmd<boolean>("image_delete", { locationId, assetPath }, onOk, onErr);
-}
-
-export function imageList(...[locationId, onOk, onErr]: ImageListParams<ImageAsset[]>): Cmd {
-  return invokeCmd<ImageAsset[]>("image_list", { locationId }, onOk, onErr);
 }
 
 export function assetResolve(...[locationId, docRelPath, assetPath, onOk, onErr]: AssetResolveParams<string>): Cmd {