+++ src/ports/types.ts
@@ -332,11 +332,14 @@ export type PostGetMarkdownParams = Parameters<
 >;
 
 export type ImageImportParams<T> = Parameters<
-  (locationId: LocationId, sourcePath: string, onOk: SuccessCallback<T>, onErr: ErrorCallback) => void
+  (
+    locationId: LocationId,
+    sourcePath: string,
+    targetDir: string,
+    onOk: SuccessCallback<T>,
+    onErr: ErrorCallback,
+  ) => void
 >;
 export type ImageDeleteParams<T> = Parameters<
   (locationId: LocationId, assetPath: string, onOk: SuccessCallback<T>, onErr: ErrorCallback) => void
->;
-export type ImageListParams<T> = Parameters<
-  (locationId: LocationId, onOk: SuccessCallback<T>, onErr: ErrorCallback) => void
 >;